requirements.txt
alabaster==0.7.12
awscli==1.27.38
Babel==2.11.0
botocore==1.29.38
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
contourpy==1.0.6
coverage==7.0.3
cycler==0.11.0
docutils==0.16
flake8==6.0.0
fonttools==4.38.0
idna==3.4
imagesize==1.4.1
intel-openmp==2022.2.1
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mccabe==0.7.0
mkl==2022.2.1
# mkl-fft==1.3.1
# mkl-random==1.2.2
# mkl-service==2.4.0
numpy==1.23.5
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pip==22.3.1
pyaml==21.10.1
pyasn1==0.4.8
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==5.4.1
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
seaborn==0.12.2
setuptools==65.5.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
src==0.1.0
tbb==2021.7.1
torch==1.13.1
typing_extensions==4.4.0
urllib3==1.26.13
wget==3.2
wheel==0.37.1
wincertstore==0.2
